+++ package.json
@@ -22,8 +22,8 @@
     "build:pl": "NODE_ENV='production' webpack --config ./apps/pl/webpack.pl.js",
     "dev:drupal": "NODE_ENV='development' webpack --watch --config ./apps/drupal/webpack.drupal.js",
     "build:drupal": "NODE_ENV='production' webpack --config ./apps/drupal/webpack.drupal.js",
-    "dev:grav": "NODE_ENV='development' webpack --watch --progress --config ./apps/grav/webpack.grav.dev.js",
-    "build:grav": "NODE_ENV='production' webpack --config ./apps/grav/webpack.grav.prod.js",
+    "dev:grav": "NODE_ENV='development' webpack --watch --progress --config ./apps/grav/webpack.grav.js",
+    "build:grav": "NODE_ENV='production' webpack --config ./apps/grav/webpack.grav.js",
     "lint:js": "eslint ./",
     "lint:scss": "stylelint '**/*.{css,scss}'",
     "lint": "npm run lint:js; npm run lint:scss",
